List: Use a small check mark instead of "Done".
As in android-galaxyzoo:
https://github.com/murraycu/android-galaxyzoo/blob/master/app/src/main/res/drawable-xxxhdpi/ic_small_accept.png
(which has the same license.)

However, the icon sizes in the Images assets catalog are too big. I need to
find some way to automatically generate them.

I tried using Unicode character U2714 in a UILabel but I could not
change its color from black to white. This seems to be a know problem,
though some people could get it to work with some fonts:
http://stackoverflow.com/questions/9844934/check-mark-character-prevents-text-color-with-uicolor