ZooniverseClient: generateAuthorizationHeader: Use the correct (UTF8) encoding.
With this, the server reports a successful upload, but I don't see it yet
the uploads yet on my profile page:
http://www.galaxyzoo.org/#/profile/recents/1
Something else must be wrong. We already know that the server can report
success when the apiKey is wrong, for instance:
https://github.com/zooniverse/Galaxy-Zoo/issues/184
However, in this case our Authentication header does seem to be
exactly the same as what android-galaxyzoo uses for the same login,
and that still works.