ClassifyViewController: prepareForSegue: Add constraints in code.
So make the Child Containers have the same size as the main view of
their embedded UIViewControllers.

This lets us remove the hard-coded vertical spacing from the storyboard.
It leaves a warning about the vertical sizes being ambigous, because
the storyboard doesn't know about our code, so it would still be nice
to find a way to specify this in the storyboard instead of in code.