iOS7: Add pre-iOS 8 launch image assets to support iOS 7.
iOS8 uses the LaunchScreen.xib but iOS 7 needs actual images to
fake the appearance during startup.
The 2x iPhone Portrait iOS 7, 8 image is a screenshot on the iPhone 4S simulator,
and is therefore correctly 640x960, as specified here:
https://developer.apple.com/library/ios/documentation/UserExperience/Conceptual/MobileHIG/IconMatrix.html#//apple_ref/doc/uid/TP40006556-CH27-SW2
for "iPhone 4s (@2x)"
The Retina 4 iPhone Portrait iOS 7, 8 image is a screenshot on the iPhone 5 simulator,
and is therefore 640x1136, which is apparently correct,
though the Apple documentation is very unclear about this.