Classify: Change Log In to Log Out when we are logged in.
Change the toolbar item's text to Log Out when we are logged in,
for instance after a successful log in. Handle a click on that item,
while logged in, as a log out, wiping the username and api key (auth).
Log out isn't particularly useful but it gives the user some clue
that they are logged in.
This should help with this issue:
https://github.com/murraycu/ios-galaxyzoo/issues/14