iPad: Classify: Arrange image and questions/answers horizontally.
With each one taking up half of the width of the screen.
This happens on iPad for both landscape and orientation. It doesn't look
like Apple want us to arrange layouts differently on iPad landscape
than portrait, or at least they don't make that possible with their
standard "Size Classes" in XCode's storyboard.